+++ packages/veui/demo/cases/CheckboxGroup.vue
@@ -1,16 +1,16 @@
 <template>
   <article>
-    <h1><code>&lt;veui-checkboxgroup&gt;</code></h1>
+    <h1><code>&lt;veui-checkbox-group&gt;</code></h1>
     <p>
-      <veui-checkboxgroup :items="data1" v-model="picked1">
+      <veui-checkbox-group :items="data1" v-model="picked1">
         <template scope="props">{{ props.label }}</template>
-      </veui-checkboxgroup>
+      </veui-checkbox-group>
     </p>
     <p>value: {{ picked1 }}</p>
     <p>
-      <veui-checkboxgroup :items="data2" ui="small" v-model="picked2">
+      <veui-checkbox-group :items="data2" ui="small" v-model="picked2">
         <template scope="props">{{ props.label }}</template>
-      </veui-checkboxgroup>
+      </veui-checkbox-group>
     </p>
     <p>value: {{ picked2 }}</p>
   </article>
@@ -20,9 +20,9 @@
 import { CheckboxGroup } from 'veui'
 
 export default {
-  name: 'checkboxgroup-demo',
+  name: 'checkbox-group-demo',
   components: {
-    'veui-checkboxgroup': CheckboxGroup
+    'veui-checkbox-group': CheckboxGroup
   },
   data () {
     return {