+++ packages/veui/src/components/PromptBox.vue
@@ -1,5 +1,5 @@
 <template>
-<veui-dialog class="veui-promptbox"
+<veui-dialog class="veui-prompt-box"
   :ui="ui"
   :open.sync="localOpen"
   :priority="priority"
@@ -8,8 +8,8 @@
   @ok="$emit('ok')"
   @cancel="$emit('cancel')">
   <template slot="title"><slot name="title">{{ title }}</slot></template>
-  <p class="veui-promptbox-info">{{ content }}</p>
-  <veui-input v-model="localValue" class="veui-promptbox-input"></veui-input>
+  <p class="veui-prompt-box-info">{{ content }}</p>
+  <veui-input v-model="localValue" class="veui-prompt-box-input"></veui-input>
 </veui-dialog>
 </template>
 
@@ -24,7 +24,7 @@ config.defaults({
 })
 
 export default {
-  name: 'veui-promptbox',
+  name: 'veui-prompt-box',
   components: {
     'veui-input': Input,
     'veui-dialog': Dialog