+++ packages/veui/src/components/Table/_TableHead.js
@@ -15,6 +15,7 @@ export default {
     ...table.mapTableData(
       'data',
       'selectable',
+      'selectMode',
       'selectStatus',
       { realColumns: 'columns' }
     )
@@ -25,15 +26,20 @@ export default {
         <tr>
           {
             this.selectable
-              ? <th><div class="veui-table-cell"><veui-checkbox checked={this.selectStatus !== 'none'}
-                disabled={!this.data.length}
-                indeterminate={this.selectStatus === 'partial'}
-                onChange={checked => { this.table.select(checked) }}/></div></th>
-              : ''
+              ? <th><div class="veui-table-cell">
+                {
+                  this.selectMode === 'multiple'
+                    ? <veui-checkbox checked={this.selectStatus !== 'none'}
+                        disabled={!this.data.length}
+                        indeterminate={this.selectStatus === 'partial'}
+                        onChange={checked => { this.table.select(checked) }}/>
+                    : null
+                }</div></th>
+              : null
           }
           {
             this.columns.map(col => (
-              <th class={col.align ? `veui-table-column-${col.align}` : ''}>
+              <th class={col.align ? `veui-table-column-${col.align}` : null}>
                 <div class="veui-table-cell">{col.renderHead.call(this._renderProxy, { col })}</div>
                 {
                   col.sortable
@@ -41,7 +47,7 @@ export default {
                         order={this.table.orderBy === col.field ? this.table.order : false}
                         onSort={order => { this.$emit('sort', col.field, order) }}>
                       </veui-table-sorter>
-                    : ''
+                    : null
                 }
               </th>
             ))